pyproject.toml
[tool.poetry]
name = "yatube"
version = "0.1.0"
description = ""
authors = ["Evan <evan.vilagov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
djangorestframework = "^3.12.4"
sorl-thumbnail = "^12.7.0"
gunicorn = "^20.1.0"
python-dotenv = "^0.17.1"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
